Update service to use options objects
This is a breaking change, but it's not really that hard to migrate to.

This commit also changes the ifs to only check for undefined - when value is "" `if (value)` would return false, but since the user already passed the empty string they probably want it there.

It's kinda annoying having to pass (...undefined, undefined) before getting to arguments like interaction, so this should fix it.